[hypre]: fix some coverity warnings
Create a private `hypre_linear_solver.h` header file.
