+++ plugins/hypre/src/hypre_linear_solver.cpp
@@ -16,20 +16,11 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include "hypre_matrix.h"
-#include "hypre_vector.h"
+#include "hypre_linear_solver.h"
+
 
 #include <HYPRE_parcsr_ls.h>
 #include <HYPRE_parcsr_mv.h>
-
-#include <arccore/message_passing_mpi/MpiMessagePassingMng.h>
-
-#include <alien/core/backend/LinearSolverT.h>
-#include <alien/expression/solver/SolverStater.h>
-
-#include <alien/hypre/backend.h>
-#include <alien/hypre/export.h>
-#include <alien/hypre/options.h>
 
 namespace Alien
 {
@@ -40,57 +31,6 @@ template class ALIEN_HYPRE_EXPORT LinearSolver<BackEnd::tag::hypre>;
 
 namespace Alien::Hypre
 {
-class InternalLinearSolver : public IInternalLinearSolver<Matrix, Vector>
-, public ObjectWithTrace
-{
- public:
-  typedef SolverStatus Status;
-
-  InternalLinearSolver();
-
-  InternalLinearSolver(const Options& options);
-
-  virtual ~InternalLinearSolver() {}
-
- public:
-  // Nothing to do
-  void updateParallelMng(Arccore::MessagePassing::IMessagePassingMng* pm) {}
-
-  bool solve(const Matrix& A, const Vector& b, Vector& x);
-
-  bool hasParallelSupport() const { return true; }
-
-  //! Etat du solveur
-  const Status& getStatus() const;
-
-  const SolverStat& getSolverStat() const { return m_stat; }
-
-  std::shared_ptr<ILinearAlgebra> algebra() const;
-
- private:
-  Status m_status;
-
-  Arccore::Real m_init_time;
-  Arccore::Real m_total_solve_time;
-  Arccore::Integer m_solve_num;
-  Arccore::Integer m_total_iter_num;
-
-  SolverStat m_stat;
-  Options m_options;
-
- private:
-  void checkError(const Arccore::String& msg, int ierr, int skipError = 0) const;
-};
-
-InternalLinearSolver::InternalLinearSolver()
-{
-}
-
-InternalLinearSolver::InternalLinearSolver(const Options& options)
-: m_options(options)
-{
-}
-
 void InternalLinearSolver::checkError(
 const Arccore::String& msg, int ierr, int skipError) const
 {
